+++ packages/orq-rc/src/models/operations/fileupload.ts
@@ -245,7 +245,7 @@ export const FileUploadResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-02-27T12:53:44.988Z",
+    "2025-02-27T17:20:06.978Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -279,7 +279,7 @@ export const FileUploadResponseBody$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-02-27T12:53:44.988Z"))
+  created: z.date().default(() => new Date("2025-02-27T17:20:06.978Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {