+++ packages/orq-rc/RELEASES.md
@@ -599,3 +599,13 @@ Based on:
 - [typescript v3.2.0-rc.20] packages/orq-rc
 ### Releases
 - [NPM v3.2.0-rc.20] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.20 - packages/orq-rc
+
+## 2025-02-27 17:21:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.505.2 (2.531.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.2.0-rc.21] packages/orq-rc
+### Releases
+- [NPM v3.2.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.21 - packages/orq-rc